requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
pycparser==2.19
PySocks==1.7.0
python-telegram-bot==12.0.0b1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
tornado==6.0.3
urllib3==1.25.3
